spent like 9 hours getting chunk code to work decently (have yet to attempt rendering them)

+++ utils.h
@@ -1,10 +1,14 @@
 #pragma once
 #include <SDL.h>
+#include "Renderwindow.hpp"
+
 
 namespace utils {
 	inline float hireTimeInSeconds() {
-		float t = SDL_GetTicks();
+		float t = (float)SDL_GetTicks();
 		t *= 0.001f;
 		return t;
 	}
+
 }
+
